wraped mocha with more convenient functions
